HSD #15015507511: socfpga: cyclone5: Changing CONFIG_HPS to CFG_HPS
This commit is to change in .py script of the cv_bsp_generator to change
all namespace of CONFIG_HPS to CFG_HPS. This is to follow the changes
from the community starting from U-boot v2023.01. Thus it will updated
with correct name format when run the cv_bsp_generator.py script.

Signed-off-by: Muhammad Hazim Izzat Zamri <muhammad.hazim.izzat.zamri@intel.com>